refactor hooks api partially
